feat: Project restructure
