+++ src/runtime/internal/job_scheduler.rs
@@ -1,15 +1,15 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2026 Eduard Smet */
 
-use crate::plugins::{
-    discord_bot::plugin::{
+use crate::runtime::{
+    internal::InternalRuntime,
+    plugins::wbps::plugin::{
         job_scheduler_import_functions::Host as JobSchedulerImportFunctionsHost,
         job_scheduler_import_types::{
             Host as JobSchedulerImportTypesHost, JobSchedulerRegistrations,
             JobSchedulerRegistrationsResult, SupportedJobSchedulerRegistrations,
         },
     },
-    runtime::internal::InternalRuntime,
 };
 
 impl JobSchedulerImportTypesHost for InternalRuntime {}
